+++ codegen/templates/code/controller/multicluster_reconcilers.gotmpl
@@ -1,0 +1,2 @@
+// Definitions for the Kubernetes Controllers
+package controller
